api/requirements.txt
fastapi
pydantic
pinecone==4.0.0
python-dotenv==1.0.1
uvicorn